Move bundle simulation bank tests
